+++ packages/create-app/templates/js/src/app.js
@@ -1,0 +1,10 @@
+import { html } from "@inglorious/web"
+
+export const app = {
+  render(api) {
+    return html`<h1>
+      ${api.render("message1")}, ${api.render("message2")},
+      ${api.render("message3")}!
+    </h1>`
+  },
+}