+++ packages/create-app/templates/minimal/src/store/middlewares.js
@@ -1,0 +1,7 @@
+import { createDevtools } from "@inglorious/web"
+
+export const middlewares = []
+
+if (window.process.env === "development") {
+  middlewares.push(createDevtools().middleware)
+}