+++ packages/create-app/templates/ts/src/store/middlewares.ts
@@ -1,0 +1,10 @@
+import type { Middleware } from "@inglorious/store"
+import { createDevtools } from "@inglorious/web"
+
+import type { AppEntity, AppState } from "../../types"
+
+export const middlewares: Middleware<AppEntity, AppState>[] = []
+
+if (import.meta.env.DEV) {
+  middlewares.push(createDevtools<AppEntity, AppState>().middleware)
+}