+++ packages/create-app/templates/ts/vite.config.js
@@ -1,0 +1,8 @@
+import { defineConfig } from "vite"
+
+export default defineConfig({
+  // @see https://github.com/vitejs/vite/issues/1973
+  define: { "process.env": {} },
+
+  plugins: [],
+})