+++ examples/apps/react-todomvc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # todomvc
 
+## 1.1.8
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @inglorious/store@7.1.1
+
 ## 1.1.7
 
 ### Patch Changes