+++ examples/apps/web-table/CHANGELOG.md
@@ -1,5 +1,11 @@
 # lit-todomvc
 
+## 0.1.3
+
+### Patch Changes
+
+- @inglorious/web@2.1.1
+
 ## 0.1.2
 
 ### Patch Changes