+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inglorious/eslint-config
 
+## 1.1.0
+
+### Minor Changes
+
+- Add TS lint rules
+
 ## 1.0.1
 
 ### Patch Changes