+++ packages/react-store/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inglorious/react-store
 
+## 9.0.1
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @inglorious/store@7.1.1
+
 ## 9.0.0
 
 ### Patch Changes